added adb interface factory design pattern refactored code to be device probes for lower level adb shell diagnostics added gpu level 0 diagnostics.
