+++ android_devices/adb_output_parsers.py
@@ -1,0 +1,7 @@
+
+
+
+class AdbParser:
+
+    def __init__(self, **kwargs):
+        """Constructor"""
